Don't squelch error reporting when comparators fail to load.
Explicitly check whether there is a fallback left to try, before raising
an exception in the main loop. This ensures the original traceback is
also shown.

Without this, the original cause is quite mysterious, and users need to
scrutinize the code in order to figure out how to figure out what was
actually failing to be imported.